Add stricter validation of the depends_on field for drafts.
The review request draft resource passed in any review request IDs to
the queryset when attempting to get a list of valid and invalid IDs.
However, if one of the values passed in was not an integer, this would
trigger an error within Django, which would crash the API request.

The resource now pre-filters the values, querying only the ones that
resolve to integers. Anything else ends up in the list of invalid IDs
that are reported to the client.

Testing Done:
Unit tests pass.

Reviewed at https://reviews.reviewboard.org/r/10541/