Remove deprecated signatures for SCMTool.get_file and file_exists.
`SCMTool.get_file()` and `SCMTool.file_exists()` gained the addition of
a `base_commit_id` many releases ago, and when that happened we needed
to deprecate the old signature and warn if used, but continue to support
it.

This change removes support for the deprecated signatures, which is long
overdue but also helps with Python 3 compatibility (as the function we
were using to check was also deprecated in Python 3).

Testing Done:
Unit tests pass.

Reviewed at https://reviews.reviewboard.org/r/10505/