Register the Local Site url template tag on Django 1.11.
On Django 1.6, we replace the `{% url %}` template tag with our own by
adding it to the built-in list of tags through the `add_to_builtins()`
method. On newer versions, we have a much cleaner way of doing this by
listing the module as a built-in template tag library in settings.

This change makes use of the new setting, and avoids calling
`add_to_builtins()` on newer releases where it doesn't exist.

Testing Done:
Unit tests pass on Django 1.6 and 1.11.

Manually tested the site on both to ensure that `{% url %}` was
working with Local Sites and non-Local Sites.

Reviewed at https://reviews.reviewboard.org/r/10525/