Compute the Whoosh index path setting dynamically.
Whoosh had a default setting for its index directory that was computed
at module initialization time. In unit tests, this resulted in the local
directory being used and not the directory set up for unit test state.

The Whoosh settings are now computed dynamically in order to take that
setting into account. On top of this, we also now override the path
hard-coded in `settings.py` (as part of the default Haystack
configuration, which is required) to also use the temporary storage
path.

Testing Done:
Unit tests pass on Django 1.6 and 1.11.

Verified that a `search-index` directory was no longer being created
in the source directory.

Reviewed at https://reviews.reviewboard.org/r/10529/