Update the dependencies and package install for Python 3.x.
We're now supporting Python 3.6 as a minimum for the 3.x range. This is
enforced in `setup.py` and `manage.py`, with the minimum versions now
defined in `dependencies.py`.

`setup.py` now permits packaging if setting the environment variable
`RB_PY3_HAS_NO_OFFICIAL_SUPPORT=agreed`. This helps with the
development and testing of the packaging while at the same time making
it clear to anyone doing so that we are not supporting this version yet.

It also bumps the minimum Django 1.11 release to the latest 1.11.20.

Testing Done:
Tested running `develop` and `bdist_wheel` on Python 2.7, 3.6, and 3.7.

Tested that older Python versions triggered appropriate failures to run
`setup.py` and `manage.py`.

Reviewed at https://reviews.reviewboard.org/r/10530/