Add legacy aliases for the recently-removed SSH key properties.
We have some unit tests outside of Review Board that make use of the old
`key1`, `key2`, `key1_b64`, and `key2_b64` SSH keys in `SSHTestCase`.
These broke with the migration of these keys to the newer SSH key
support for unit tests, and that code needs to work with older and newer
versions of Review Board.

This change adds legacy properties that forward on to the new keys,
preserving backwards-compatibility.

Testing Done:
Unit tests pass in Review Board and our in-house tests.

Reviewed at https://reviews.reviewboard.org/r/10538/