Use new tabs component for the admin dashboard.
This change migrates the admin dashboard to use the new tabs component.
As part of this, the tabs themselves have been moved over so they
connect with the page content. This currently looks a little wonky
because the sidebar is super wide, but that will be addressed in a
future change.

Testing Done:
Flipped through the tabs and verified the appearance.

Reviewed at https://reviews.reviewboard.org/r/10563/