Make the support banner margined rather than flush.
With the new tabs look in the admin UI, the support banner looked kind
of funny. This change makes it so it has margins on all sides rather
than being flush with the top.

Testing Done:
Verified the appearance of the support banner.

Reviewed at https://reviews.reviewboard.org/r/10565/