Simplify and shrink the admin sidebar.
The sidebar in the admin site had grown pretty unwieldy, including a lot
of information that really didn't need to be shown all of the time. This
change simplifies it a lot, eliminating the "System Information" box
completely and shrinking the width down to fit the content.

The one part of the System Information box which really still needed to
be linked was the security checklist, which has been added into the top
("System Settings") box and renamed as "Security Center".

Testing Done:
Verified the appearance of the sidebar.

Reviewed at https://reviews.reviewboard.org/r/10564/