Correctly configure the codehilite Markdown extension.
We had a configuration option set for the Markdown `codehilite`
extension that referenced it by the old name rather than the module
path. This only worked on older versions of Python Markdown, and failed
on anything modern, resulting in differences in behavior depending on
the version being used. The module-based name is always safe, so it's a
simple fix to just use that one.

Testing Done:
Unit tests pass against new versions of Python Markdown.

Reviewed at https://reviews.reviewboard.org/r/10586/