Fix an iteration ordering issue with RB.ReviewRequestActionHook.
The JavaScript `RB.ReviewRequestActionHook` had an issue with the order
of data when iterating through the list of registered hooks. What we
documented in the manual was correct, but the code implementation
wasn't. This meant that anyone following the docs would find that their
hooks weren't doing anything but crashing.

This is a simple fix that switches the ordering to be correct.

Testing Done:
Unit tests pass.

Reviewed at https://reviews.reviewboard.org/r/10588/