Add release notes for Review Board 3.0.15.
This is an important bug fix release that fixes support for Bitbucket
(due to the removal of an API). It also contains fixes for avatars
rendering and emojis, saving of access lists for repositories,
validation of the Depends On field, and more.

Testing Done:
Built the docs. Checked for build errors, bad links, and spelling
issues.

Reviewed at https://reviews.reviewboard.org/r/10587/