Further bullet-proof Bitbucket webhook payload parsing.
A recent change attempted to deal with Bitbucket payload changes that
resulted in fields sometimes not being present in the payload, depending
on the event that fired. We've since found that sometimes the keys are
present, but are null, breaking some of our parsing. This change further
improves it, handling both missing keys and null values for those keys,
and falling back to defaults in both cases.

Testing Done:
Unit tests pass.

Reviewed at https://reviews.reviewboard.org/r/10595/