Fix admin widget adder modals.
The widget adder dialogs were using jQuery UI's `$.dialog`, which at
some point got screwed up. This resulted in ugly dialogs with broken
CSS, a modal background which wasn't correct (was shown over the widget
itself and didn't scroll with the page), and it didn't fit visually with
the rest of the UI.

This change ports it over to use our `$.modalBox`, which is what
everything else uses for modals. This makes it fit in with our UI, and
just generally function correctly.

Testing Done:
Added and removed both small and large widgets and saw that the dialog
displayed correctly.

Reviewed at https://reviews.reviewboard.org/r/10609/