Fix unit tests when hg is uninstalled or matches a different Python.
We had a unit test that was checking for the `mercurial` module to be
loaded before attempting to create a Mercurial-based repository. This
check wasn't correct, because we don't care if the module is accessible.
We only care if the `hg` binary is.

Another unit test wasn't performing a check at all, erroring out instead
when trying to create the repository.

This change fixes both issues by checking for the `hg` binary to be
available before running the test.

Testing Done:
Unit tests pass with and without the `hg` binary (and without access to
the `mercurial` module).

Reviewed at https://reviews.reviewboard.org/r/10621/