Fix regressions in the repository form ES6 transition.
When the repository form's code was moved to ES6, there were a couple of
regressions that broke the form's functionality.

First, the page was crashing when trying to set the help text. It was
attempting to use a variable that was no longer defined. We now just
pass the value originally set for that variable into the function that
needed it.

Second, a typo in a selector resulted in SCMTool-specific rows not being
hidden properly, causing some confusion when trying to configure a
hosting service.

This brings the repository form back to working order.

Testing Done:
Created some repositories and switched between hosting services and
repository types. Everything appeared to behave as expected.

Reviewed at https://reviews.reviewboard.org/r/10626/