Quote values going into URLs for RB Gateway.
To help form valid URLs, we now ensure that all values going into a RB
Gateway URL is quoted. This ensures that if there's a character in a
branch name or file path that doesn't translate properly to a URL, it
will be encoded in such a way where the server can handle it.

Testing Done:
Unit tests pass.

Reviewed at https://reviews.reviewboard.org/r/10637/