Show the hosting service auth form if it has any errors.
If editing credentials for an existing account entry for a hosting
service, and validation fails on any of the fields, the auth form would
be hidden, hiding the fact that anything was wrong.

We now check whether there are any errors to show, and ensure we always
show the form if any are found.

Testing Done:
Hit this while testing a change to Bitbucket. Verified that the form
would always show up when triggering a validation failure, but was
otherwise hidden if the fields passed validation.

Reviewed at https://reviews.reviewboard.org/r/10640/