Update to the latest CodeMirror release.
CodeMirror has had a bunch of releases since we last updated. Most of
these deal with individual modes, but there have been some good bugfixes
that we want to pick up. This change updates us to the latest released
version, 5.48.2. As part of this I've updated the instructions for
building the file to cope with changes in their tree (the removal of the
`compress` script) and switched us to bundling an un-minified file,
which makes things easier for debugging.

Testing Done:
Smoke tested various text editors and different features of the editor.

Reviewed at https://reviews.reviewboard.org/r/10647/