Load djblets-widgets on all pages, not just admin.
We moved the related object selector implementation into djblets, and as
part of that change the new djblets-widgets bundle was only loaded for
admin pages. Unfortunately this broke rbintegrations features which
relied on selectize for some special fields. This change loads the
bundle on all pages, not just those in admin.

Testing Done:
Was able to use rbintegrations fields based on selectize.

Reviewed at https://reviews.reviewboard.org/r/10650/