Perform incremental search reindexing of users based on the last_login field.
When performing an incremental reindexing of users, Haystack emits
warnings that it has no way of determining the age of an entry for
users, since there's no sort of "last updated" field available.

We do have a workaround, in that Review Board does update the
``last_login`` field when users access the site, rather than just when
they explicitly log in. We can tie updates to that.

What this basically means is that we will re-index any active accounts
(which might have information changed), but if information is changed
with an older account, it won't be noticed without performing a full
re-index.

Testing Done:
Performing an incremental reindexing and saw that the warnings went
away, and the new information on a user was indexed.

Reviewed at https://reviews.reviewboard.org/r/10645/