Remove deprecated support for unset hosting service IDs.
Review Board 3.0 is gaining support for setting an explicit hosting
service ID, which we previously didn't correctly support when using a
`HostingServiceHook`. It was defaulting to using the service's name,
which wasn't a valid ID. To fix this, it automatically slugified the
name, turning it into an ID, but this is meant to be temporary,
deprecated behavior for 3.0.

This change removes that behavior, replacing it instead with an error
that's raised if the ID is missing.

It also removes some unused imports from previous deprecation removals.

Testing Done:
Unit tests pass.

Reviewed at https://reviews.reviewboard.org/r/10652/